♻️ Divide files
